requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
psycopg2-binary==2.9.9 
python-dotenv==1.0.1
redis==5.0.1
gunicorn==21.2.0
Werkzeug==3.0.1
SQLAlchemy==2.0.28
email-validator==2.1.0
psutil==5.9.8
